# PERF: ArrowExtensionArray.__iter__