--- a/pandas/core/arrays/arrow/array.py
+++ b/pandas/core/arrays/arrow/array.py
@@ -10,6 +10,7 @@ import numpy as np
 
 from pandas._typing import (
     Dtype,
+    Iterator,
     PositionalIndexer,
     SortKind,
     TakeIndexer,
@@ -329,6 +330,18 @@ class ArrowExtensionArray(OpsMixin, ExtensionArray):
                 return self._dtype.na_value
             else:
                 return scalar
+
+    def __iter__(self) -> Iterator[Any]:
+        """
+        Iterate over elements of the array.
+        """
+        na_value = self._dtype.na_value
+        for value in self._data:
+            val = value.as_py()
+            if val is None:
+                yield na_value
+            else:
+                yield val
 
     def __arrow_array__(self, type=None):
         """Convert myself to a pyarrow ChunkedArray."""